+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Clear ``ON CONFLICT``, ``UPDATE``, and ``WHERE`` clauses in ``INSERT`` statements.
+
 1.1.0 (2025-04-03)
 ------------------